requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
